refactor(core): 合并 entities/ 和 types/ 为 models/
dataclass 和 TypedDict 是实现细节，不是业务概念。
合并后找类型只需看一个目录，消除 entities vs types 的选择负担。
