+++ src/danmaku_sender/core/engines/editor_session.py
@@ -3,10 +3,10 @@ import logging
 import colorsys
 from typing import Any, Callable
 
-from ..state import ValidationConfig
-from ..entities.danmaku import Danmaku
-from ..types.editor_types import EditorItem, EditorField, AtomicChange, ViewItem, InsertPosition
-from ..services.danmaku_validator import validate_danmaku_list
+from danmaku_sender.core.state import ValidationConfig
+from danmaku_sender.core.models.danmaku import Danmaku
+from danmaku_sender.core.models.editor_types import EditorItem, EditorField, AtomicChange, ViewItem, InsertPosition
+from danmaku_sender.core.services.danmaku_validator import validate_danmaku_list
 
 
 class EditorSession:
