+++ src/danmaku_sender/core/engines/sender/executor.py
@@ -1,12 +1,12 @@
 import logging
 from threading import Event
 
-from ....api.bili_api_client import BiliApiClient
-from ...entities.danmaku import Danmaku
-from ...types.common import VideoTarget
-from ...types.result import DanmakuSendResult
-from ...exceptions.exceptions import BiliApiError, BiliNetworkError
-from ...exceptions.api_errors import BiliDmErrorCode
+from danmaku_sender.api.bili_api_client import BiliApiClient
+from danmaku_sender.core.models.danmaku import Danmaku
+from danmaku_sender.core.models.common import VideoTarget
+from danmaku_sender.core.models.result import DanmakuSendResult
+from danmaku_sender.core.exceptions.exceptions import BiliApiError, BiliNetworkError
+from danmaku_sender.core.exceptions.api_errors import BiliDmErrorCode
 
 
 class DanmakuExecutor:
