+++ src/danmaku_sender/core/models/result.py
