+++ src/danmaku_sender/core/services/video_fetcher.py
@@ -1,8 +1,7 @@
 import logging
 
-from ..entities.video import VideoInfo, VideoPart
-
-from ...api.bili_api_client import BiliApiClient
+from danmaku_sender.api.bili_api_client import BiliApiClient
+from danmaku_sender.core.models.video import VideoInfo, VideoPart
 
 
 class VideoFetcher:
