+++ src/danmaku_sender/ui/controllers/auth_controller.py
@@ -3,11 +3,10 @@ import threading
 
 from PySide6.QtCore import QObject, Signal, Slot
 
-from ..framework.concurrency import WorkerThread, PoolTask
-
-from ...api.bili_api_client import BiliApiClient
-from ...core.state import ApiAuthConfig
-from ...core.entities.user import UserProfile
+from danmaku_sender.api.bili_api_client import BiliApiClient
+from danmaku_sender.core.state import ApiAuthConfig
+from danmaku_sender.core.models.user import UserProfile
+from danmaku_sender.ui.framework.concurrency import WorkerThread, PoolTask
 
 
 logger = logging.getLogger("App.System.Auth")
