+++ src/danmaku_sender/ui/controllers/monitor_controller.py
@@ -5,11 +5,11 @@ from PySide6.QtCore import QObject, Signal, Slot
 
 from ..framework.concurrency import WorkerThread
 
-from ...core.types.common import VideoTarget
-from ...core.state import ApiAuthConfig, MonitorConfig
-from ...core.engines.bili_monitor import BiliDanmakuMonitor
-from ...api.bili_api_client import BiliApiClient
-from ...utils.system_utils import KeepSystemAwake
+from danmaku_sender.core.models.common import VideoTarget
+from danmaku_sender.core.state import ApiAuthConfig, MonitorConfig
+from danmaku_sender.core.engines.bili_monitor import BiliDanmakuMonitor
+from danmaku_sender.api.bili_api_client import BiliApiClient
+from danmaku_sender.utils.system_utils import KeepSystemAwake
 
 
 logger = logging.getLogger("App.System.MonitorController")
