+++ src/danmaku_sender/ui/controllers/sender_controller.py
@@ -9,9 +9,9 @@ from danmaku_sender.ui.framework.concurrency import WorkerThread, PoolTask
 from danmaku_sender.core.database.history_manager import HistoryManager
 from danmaku_sender.core.engines.sender import DanmakuScheduler, DanmakuExecutor, SendingContext, SendJob
 from danmaku_sender.core.engines.sender.delay_manager import DelayManager
-from danmaku_sender.core.entities.danmaku import Danmaku
-from danmaku_sender.core.types.result import DanmakuSendResult
-from danmaku_sender.core.types.common import VideoTarget, UnsentDanmakusRecord
+from danmaku_sender.core.models.danmaku import Danmaku
+from danmaku_sender.core.models.result import DanmakuSendResult
+from danmaku_sender.core.models.common import VideoTarget, UnsentDanmakusRecord
 from danmaku_sender.core.state import ApiAuthConfig, SenderConfig, AppState
 from danmaku_sender.core.services.danmaku_parser import DanmakuParser
 from danmaku_sender.core.services.danmaku_exporter import create_xml_from_danmakus
